Fused GenerateResource and StoreResource because Camunda will persist the generated Binary content between tasks, causing a OutOfMemory exception if the Binary content is too large to be persisted
