+++ src/main/java/dev/dsf/bpe/service/GenerateAndStoreResource.java
@@ -1,10 +1,12 @@
 package dev.dsf.bpe.service;
 
+import java.io.ByteArrayInputStream;
 import java.util.Base64;
 import java.util.Random;
 
 import org.camunda.bpm.engine.delegate.BpmnError;
 import org.camunda.bpm.engine.delegate.DelegateExecution;
+import org.hl7.fhir.r4.model.IdType;
 
 import dev.dsf.bpe.ConstantsPing;
 import dev.dsf.bpe.util.logging.PingPongLogger;
@@ -12,11 +14,11 @@ import dev.dsf.bpe.v1.ProcessPluginApi;
 import dev.dsf.bpe.v1.activity.AbstractServiceDelegate;
 import dev.dsf.bpe.v1.variables.Variables;
 
-public class GenerateResource extends AbstractServiceDelegate
+public class GenerateAndStoreResource extends AbstractServiceDelegate
 {
 	private final int maxUploadSizeBytes;
 
-	public GenerateResource(ProcessPluginApi api, int maxUploadSizeBytes)
+	public GenerateAndStoreResource(ProcessPluginApi api, int maxUploadSizeBytes)
 	{
 		super(api);
 		this.maxUploadSizeBytes = maxUploadSizeBytes;
@@ -25,13 +27,23 @@ public class GenerateResource extends AbstractServiceDelegate
 	@Override
 	protected void doExecute(DelegateExecution delegateExecution, Variables variables) throws BpmnError, Exception
 	{
-		PingPongLogger logger = new PingPongLogger(GenerateResource.class, variables.getStartTask());
+		PingPongLogger logger = new PingPongLogger(GenerateAndStoreResource.class, variables.getStartTask());
 		logger.debug("Generating resource...");
 		int downloadResourceSizeBytes = getDownloadResourceSize(variables);
 
 		variables.setByteArray(ConstantsPing.BPMN_EXECUTION_VARIABLE_DOWNLOAD_RESOURCE,
 				generateRandomBinaryContent(downloadResourceSizeBytes, logger));
 		logger.debug("Generated resource.");
+		logger.debug("Storing binary resource for download...");
+
+		IdType downloadResource = storeBinary(
+				variables.getByteArray(ConstantsPing.BPMN_EXECUTION_VARIABLE_DOWNLOAD_RESOURCE));
+
+		String reference = downloadResource.getValueAsString();
+
+		variables.setString(ConstantsPing.BPMN_EXECUTION_VARIABLE_DOWNLOAD_RESOURCE_REFERENCE, reference);
+
+		logger.debug("Stored binary resource for download");
 	}
 
 	private byte[] generateRandomBinaryContent(int desiredSizeBytes, PingPongLogger logger)
@@ -60,5 +72,11 @@ public class GenerateResource extends AbstractServiceDelegate
 	private int getDownloadResourceSize(Variables variables)
 	{
 		return variables.getInteger(ConstantsPing.BPMN_EXECUTION_VARIABLE_DOWNLOAD_RESOURCE_SIZE_BYTES);
+	}
+
+	private IdType storeBinary(byte[] downloadResourceContent)
+	{
+		return api.getFhirWebserviceClientProvider().getLocalWebserviceClient().withMinimalReturn()
+				.createBinary(new ByteArrayInputStream(downloadResourceContent), ConstantsPing.DOWNLOAD_RESOURCE_MIME_TYPE, api.getOrganizationProvider().getLocalOrganization().get().getIdElement().getIdPart());
 	}
 }
